started working on code to transition codebase to joml

+++ engine/src/main/java/org/terasology/input/lwjgl/LwjglMouseDevice.java
@@ -25,7 +25,7 @@ import org.terasology.input.ButtonState;
 import org.terasology.input.InputType;
 import org.terasology.input.device.MouseAction;
 import org.terasology.input.device.MouseDevice;
-import org.terasology.math.geom.Vector2i;
+import org.joml.Vector2i;
 
 import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
@@ -46,7 +46,7 @@ public class LwjglMouseDevice implements MouseDevice, PropertyChangeListener {
 
     @Override
     public Vector2i getPosition() {
-        return new Vector2i(Mouse.getX() / this.uiScale, (Display.getHeight() - Mouse.getY()) / this.uiScale);
+        return new Vector2i((int)(Mouse.getX() / this.uiScale), (int)((Display.getHeight() - Mouse.getY()) / this.uiScale));
     }
 
     @Override
