+++ engine/src/main/java/org/terasology/logic/selection/LocalPlayerBlockSelectionByItemSystem.java
@@ -26,8 +26,8 @@ import org.terasology.logic.common.ActivateEvent;
 import org.terasology.logic.location.LocationComponent;
 import org.terasology.logic.players.LocalPlayer;
 import org.terasology.math.Region3i;
-import org.terasology.math.geom.Vector3f;
-import org.terasology.math.geom.Vector3i;
+import org.joml.Vector3f;
+import org.joml.Vector3i;
 import org.terasology.registry.In;
 import org.terasology.world.selection.BlockSelectionComponent;
 import org.terasology.world.selection.event.SetBlockSelectionEndingPointEvent;
@@ -98,7 +98,7 @@ public class LocalPlayerBlockSelectionByItemSystem extends BaseComponentSystem {
         if (blockSelectionComponent.isMovable) {
             Region3i region = blockSelectionComponent.currentSelection;
 
-            blockSelectionComponent.currentSelection = Region3i.createFromCenterExtents(new Vector3i(targetLocation.x, targetLocation.y, targetLocation.z),
+            blockSelectionComponent.currentSelection = Region3i.createFromCenterExtents(new Vector3i((int)targetLocation.x, (int)targetLocation.y, (int)targetLocation.z),
                     new Vector3i(region.sizeX()/2, 0, region.sizeZ()/2));
             blockSelectionComponentEntity.saveComponent(blockSelectionComponent);
