+++ engine/src/main/java/org/terasology/world/selection/BlockSelectionSystem.java
@@ -22,8 +22,8 @@ import org.terasology.entitySystem.systems.RegisterMode;
 import org.terasology.entitySystem.systems.RegisterSystem;
 import org.terasology.logic.location.LocationComponent;
 import org.terasology.math.Region3i;
-import org.terasology.math.geom.Vector3f;
-import org.terasology.math.geom.Vector3i;
+import org.joml.Vector3f;
+import org.joml.Vector3i;
 import org.terasology.world.selection.event.SetBlockSelectionEndingPointEvent;
 import org.terasology.world.selection.event.SetBlockSelectionStartingPointEvent;
 
@@ -51,7 +51,7 @@ public class BlockSelectionSystem extends BaseComponentSystem {
 
         Vector3f worldPosition = locationComponent.getWorldPosition();
 
-        Vector3i startPosition = new Vector3i(worldPosition.x, worldPosition.y, worldPosition.z);
+        Vector3i startPosition = new Vector3i((int) worldPosition.x, (int) worldPosition.y, (int) worldPosition.z);
         blockSelectionComponent.startPosition = startPosition;
         Vector3i endPosition = startPosition;
         blockSelectionComponent.currentSelection = Region3i.createBounded(startPosition, endPosition);
@@ -74,7 +74,7 @@ public class BlockSelectionSystem extends BaseComponentSystem {
 
         Vector3f worldPosition = locationComponent.getWorldPosition();
 
-        Vector3i endPosition = new Vector3i(worldPosition.x, worldPosition.y, worldPosition.z);
+        Vector3i endPosition = new Vector3i((int) worldPosition.x, (int) worldPosition.y, (int) worldPosition.z);
         Vector3i startPosition = blockSelectionComponent.startPosition;
         if (null == startPosition) {
             startPosition = endPosition;
